verl/requirements.txt
# requirements.txt records the full set of dependencies for development
torch
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict<=0.6.2
torchdata
transformers
vllm==0.8.4
wandb
packaging>=20.0
uvicorn
fastapi
sacrebleu
unbabel-comet
